Rewrite updates service as a saga
