+++ src/actions.js
@@ -8,6 +8,7 @@ export const CERTIFICATES_CLEARED = 'certificates/cleared';
 export const DEEP_LINK_TRIGGERED = 'deep-link/triggered';
 export const LOADING_ERROR_VIEW_RELOAD_SERVER_CLICKED = 'loading-error-view/reload-server-clicked';
 export const MAIN_WINDOW_STATE_CHANGED = 'main-window/state-changed';
+export const MAIN_WINDOW_INSTALL_UPDATE_CLICKED = 'main-window/install-update-clicked';
 export const MENU_BAR_ABOUT_CLICKED = 'menu-bar/about-clicked';
 export const MENU_BAR_ADD_NEW_SERVER_CLICKED = 'menu-bar/add-new-server-clicked';
 export const MENU_BAR_CLEAR_TRUSTED_CERTIFICATES_CLICKED = 'menu-bar/clear-trusted-certificates-clicked';
@@ -57,6 +58,7 @@ export const UPDATES_ERROR_THROWN = 'updates/error-thrown';
 export const UPDATES_NEW_VERSION_AVAILABLE = 'updates/new-version-available';
 export const UPDATES_NEW_VERSION_NOT_AVAILABLE = 'updates/new-version-not-available';
 export const UPDATES_READY = 'updates/ready';
+export const UPDATES_UPDATE_DOWNLOADED = 'updates/update-downloaded';
 export const WEBVIEW_CERTIFICATE_DENIED = 'webview/certificate-denied';
 export const WEBVIEW_CERTIFICATE_TRUSTED = 'webview/certificate-trusted';
 export const WEBVIEW_DID_NAVIGATE = 'webview/did-navigate';
