+++ src/components/UpdatesProvider/index.js
@@ -1,0 +1,8 @@
+import { updatesSaga } from './sagas';
+import { useSaga } from '../SagaMiddlewareProvider';
+
+export function UpdatesProvider({ children }) {
+	useSaga(updatesSaga, []);
+
+	return children;
+}
